# documentation/swagger.yaml
...
swagger: '2.0'
info:
  title: 1ClickTrips API
  description: Business travel at its best
  version: "1.0.0"
# the domain of the service
host: api.1clicktrips.com
...
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
...
  /search/events:
    get:
      summary: Event List
      description: |
        The events endpoint will deliver a list of events to the client
        that match with the given query criteria.
      tags:
        - Search
      parameters:
        - name: search_string
          in: query
          description: The search string to filter the events for
          required: true
          type: string
        - name: start_date
          in: query
          description: The start date of the trip to be planned
          required: false
          type: string
        - name: end_date
          in: query
          description: The end date of the trip to be planned
          required: false
          type: string
        - name: limitation
          in: query
          description: The maximum number of events to return
          required: false
          type: integer
          format: int32
      responses:
          200:
            description: An array of events
            schema:
              type: array
              items:
                $ref: '#/definitions/Event'
          default:
            description: Unexpected error
            schema:
              $ref: '#/definitions/Error'  
  /search/trips:
    post:
      summary: Trip Results
      description: | 
        The Trips endpoint returns a collection of trips that matches with 
        the given parameters. 
      tags:
        - Search
      parameters:
        - name: start_description
          in: body
          description: Description of start location
          required: false
          schema:
            type: string
        - name: start_latitude
          in: body
          description: Latitude component of start location
          required: true
          schema:
            type: number
            format: double
        - name: start_longitude
          in: body
          description: Longitude component of start location
          required: true
          schema:
            type: number
            format: double
        - name: end_description
          in: body
          description: Description of end location
          required: false
          schema:
            type: string
        - name: end_latitude
          in: body
          description: Latitude component of end location
          required: true
          schema:
            type: number
            format: double
        - name: end_longitude
          in: body
          description: Longitude component of end location
          required: true
          schema:
            type: number
            format: double
        - name: start_time
          in: body
          description: Time string for the local start time of the appointment
          required: true
          schema:
            type: string
            format: '2014-11-12T19:05:00'
        - name: end_time
          in: body
          description: Time string for the local end time of the appointment
          required: true
          schema:
            type: string
            format: '2014-11-12T19:05:00'
        - name: roundtrip
          in: body
          description: Roundtrip option
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: An array of trips
          schema:
            type: array
            items:
              $ref: '#/definitions/Itinerary'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /search/alternatives:
    get:
      summary: Trip Alternatives
      description: | 
        The Alternatives endpoint returns a collection of alternatives 
        for a given trip segment. 
      tags:
        - Search
      parameters:
        - name: trip_id
          in: query
          description: The trip ID of the requested trip
          required: true
          type: string
        - name: segment_id
          in: query
          description: The segment ID of the segment to be replaced
          required: true
          type: string
        - name: limitation
          in: query
          description: The maximum number of alternatives
          required: false
          type: integer
          format: int32
      responses:
          200:
            description: An array of alternatives
            schema:
              type: array
              items:
                $ref: '#/definitions/Segment' # TODO create proper format
          default:
            description: Unexpected error
            schema:
              $ref: '#/definitions/Error'          
              
  #########################################################################
  # Language API
  #########################################################################
...
definitions:
...
  Itinerary:
    properties:
      id: 
        type: string
        description: Unique identifier
      outbound:
        $ref: '#/definitions/Trip'
      inbound:
        $ref: '#/definitions/Trip'
      price:
        type: number
        format: double
      currency:
        type: string
        description: The ISO 4217 currency code
      type: 
        type: integer
        format: int32
        description: |
          Defines the trip result type
            0 = budget
            1 = time
            2 = comfort
  Trip:
    properties:
      id:
        type: string
      origin: 
        $ref: '#/definitions/Location'
      destination:
        $ref: '#/definitions/Location'
      departureTime: 
        type: string
        format: '2014-11-12T19:05:00'
      arrivalTime: 
        type: string
        format: '2014-11-12T19:05:00'
      distance: 
        type: number
        format: double
      duration:
        type: integer
        format: int32
      segments:
        type: array
        items: 
          $ref: '#/definitions/Segment'
  Location:
    properties:
      latitude:
        type: number
        format: double
      longitude:
        type: number
        format: double
  Segment: 
    properties:
      id:
        type: string
      origin: 
        $ref: '#/definitions/Location'
      destination:
        $ref: '#/definitions/Location'
      departureTime: 
        type: string
        format: '2014-11-12T19:05:00'
      arrivalTime: 
        type: string
        format: '2014-11-12T19:05:00'
      duration:
        type: number
        format: int32
        description: Duration in minutes
      type:
        type: number
        format: int32
        description: | 
          Travel option used. 
          Stay = 0, Walk = 1, Taxi = 2, Car = 4, 
          Bus = 8, Train = 16, Plane = 32
      price:
        type: number
        format: double
      currency:
        type: string
        description: The ISO 4217 currency code
...
  Event:
    required:
      - id
      - title
      - location
    properties:
      id:
        type: integer
        format: int64
      title:
        type: string
      description:
        type: string
      location:
        type: object
        $ref: '#/definitions/Location'
      tags:
        type: array
        items:
          type: string
      dates:
        type: array
        items: 
           $ref: '#/definitions/Timing'
      open:
        type: boolean
        description: |
          Indicates if the event is open, when the user planned 
          the according trip.
      url:
        type: string
      image:
        type: string
        description: Image url of the event
  Timing:
    properties:
      start:
        type: string
        description: |
          ISO 8601 time and date
          "2015-02-09T02:54:51+00:0"
      end:
        type: string
        description: |
          ISO 8601 time and date
          "2015-02-09T02:54:51+00:0"
  Error:
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string